fix(dreamer): correctness fixes from multi-agent code review (C1 critical + 11 more)
A 10-dimension adversarial code review (57 agents, each finding re-verified
against source) surfaced one critical and several high/medium correctness bugs
in the actor-critic / world-model path. This commit applies the confirmed set.

CRITICAL
- C1 lambda-return value-target off-by-one (dreamer/agent.py:_actor_critic_loss,
  rssm.imagine). imagine() returns post-step states [s_1..s_H] (start state s_0
  never scored), but lambda_return bootstraps with values[t+1] assuming a
  departure-aligned value array. The value series was V(s_1..s_{H+1}), so every
  TD target bootstrapped off V(s_{t+2}) instead of V(s_{t+1}) — a one-state-too-
  far shift that inflated EVERY target (golden check: live [46,45,43] vs correct
  [36,35,33]). This is the root cause of the session-long imag/return_mean >> real
  return and return/scale blow-up that was being mis-attributed to reward shaping.
  Fix: build a departure-aligned value array V(s_0..s_H) with V(s_0) from the
  imagination-start latent, drop the extra img_step bootstrap, and evaluate the
  critic/actor objective on departure latents feat(s_0..s_{H-1}).

HIGH
- A1 actor loss now weights each imagined step by cumprod(gamma*continues)
  (w_0=1), so steps after a predicted termination get ~zero weight instead of
  training the actor on unreachable post-crash states. Entropy + critic loss are
  weighted by the same w (Critic.loss gains an optional per-element weight).
- I1 is_first off-by-one (dreamer/env_wrapper.py): Isaac Lab auto-resets in-step,
  so the returned frame is already the next episode's first frame when is_last is
  True. The wrapper carried the PREVIOUS step's flag forward, landing is_first one
  frame late (new episode's 2nd frame) — the RSSM reset never fired on the true
  first frame and leaked the dying episode's deter/action. Now tag is_first from
  the same step's is_last.

MEDIUM
- A3 continue head trains on is_terminal (terminated only), not is_last
  (terminated|truncated). Threaded is_terminal through env_wrapper -> train_dreamer
  -> replay_buffer; cont_target = 1 - is_terminal so 20s timeouts stay
  bootstrappable. Back-compat: defaults to is_last when key absent.
- A4 actor entropy now includes the tanh log-Jacobian, so the regularizer SEES
  mean-driven saturation (the old pre-tanh-only entropy depended on log_std alone
  and was blind to it). With A5 this lets the inner mean-tanh be removed.
- A5 dropped the double-tanh on the mean. It capped both eval AND on-policy
  collection actions to ~+-0.76 (24% authority loss). act_deterministic now
  returns tanh(mean) and can command full +-1.
- A2 loss_scale_repval default 0.0 — the repval term regressed the online critic
  onto the same target_critic used for bootstrapping (redundant self-distill).
- N1 TwoHot bins built in float32 unconditionally (were logits.dtype -> bf16
  quantised the twohot soft labels, biasing reward/critic regression up to ~0.8
  nat/sample under autocast).
- R1 padded terminal frames excluded from reward/cont/decoder losses via a per-
  step mask threaded through replay_buffer; pads no longer train the heads on a
  frozen-crash-frame chain.
- R2 resume re-enters a random-action warmup window (agent._warmup_until_step)
  since the replay buffer is not checkpointed and a resumed run starts cold.
- NE1 ne_dreamer_base.yaml kl_free 1.0 -> 0.1 (the R2 remediation; 1.0 pins KL at
  >=32 nats and zeros the gradient).

CLEANUP / NITS
- Deleted dead modules: dreamer/world_model.py, dreamer/actor_critic.py,
  dreamer/utils.py (a closed cluster importing only each other; divergent
  duplicate lambda_return/Actor/twohot that a maintainer could edit expecting
  effect). CLAUDE.md updated.
- DreamerConfig: amp_dtype default float16 -> bfloat16 (matches the only configs;
  GradScaler enable logic already correctly gates on float16), docstring fixed.
- return_ema_alpha decoupled from slow_target_fraction (default 0.02 preserves
  behaviour) so the value normaliser and target-critic lag tune independently.
- LR warmup uses min(1.0, (count+1)/warmup) so the final warmup step reaches full lr.
- Removed redundant pos_embed re-init in NEDreamerTransformer.

Tests: added tests/test_actor_critic.py (lambda_return golden cases: MC,
1-step TD, termination-kills-bootstrap; end-to-end _actor_critic_loss finite;
repval-disabled + amp default guards). Full suite 20/20 passing.

+++ dreamer/distributions.py
@@ -161,17 +161,22 @@ class TwoHot:
         self.bins = logits.shape[-1]
         self.low = low
         self.high = high
-        self._bin_values = torch.linspace(low, high, self.bins, device=logits.device, dtype=logits.dtype)
+        # Bins ALWAYS float32, regardless of logits dtype. Under bf16/fp16 autocast the reward
+        # and critic logits are low-precision; building the bin lattice in that dtype quantised
+        # the twohot soft-label interpolation weights (~3 sig digits), biasing the reward/value
+        # regression by up to ~0.8 nat/sample (review N1). log_prob upcasts the logits to f32, so
+        # f32 bins keep the whole interpolation consistent.
+        self._bin_values = torch.linspace(low, high, self.bins, device=logits.device, dtype=torch.float32)
 
     def mode(self) -> torch.Tensor:
-        probs = torch.softmax(self.logits, dim=-1)
+        probs = torch.softmax(self.logits.float(), dim=-1)
         return (probs * self._bin_values).sum(-1)
 
     def mean(self) -> torch.Tensor:
         return self.mode()
 
     def log_prob(self, target: torch.Tensor) -> torch.Tensor:
-        target = target.clamp(self.low, self.high)
+        target = target.float().clamp(self.low, self.high)
         bins = self._bin_values
         idx = torch.searchsorted(bins.contiguous(), target.contiguous())
         idx = idx.clamp(1, self.bins - 1)
